requirements.txt
einops==0.3.0
numpy==1.24.4
Pillow==9.2.0
Pillow==10.3.0
torch==2.1.0
torchvision==0.16.0
transformers==4.37.0.dev0
